feat(mapped-types): add skip null properties option to partial type
With this change, you can create a partial class which disallows null values, but allows
undefined values. Previously, every class created with PartialType allowed null values for
every property, which may be undesired if you are defining the DTO for a PATCH endpoint.
If the option is not defined, the behaviour is unchanged from the previous behaviour.